minor animation code edits

+++ ORRE-v2/src/orre/sceneGraph/SimpleSceneNode.java
@@ -3,8 +3,8 @@ package orre.sceneGraph;
 import java.util.ArrayList;
 
 public class SimpleSceneNode extends EmptySceneNode{
-	protected float rotationX, rotationY, rotationZ;
-	protected float x, y, z;
+	protected double rotationX, rotationY, rotationZ;
+	protected double x, y, z;
 	protected boolean visible = true;
 	protected ArrayList<SceneNode> children = new ArrayList<SceneNode>();
 	protected float renderRadius = 0.0f;
@@ -14,28 +14,28 @@ public class SimpleSceneNode extends EmptySceneNode{
 		return this.renderRadius;
 	}
 	
-	public void translate(float x, float y, float z) 
+	public void translate(double x, double y, double z) 
 	{
 		this.x += x;
 		this.y += y;
 		this.z += z;
 	}
 
-	public void rotate(float x, float y, float z) 
+	public void rotate(double x, double y, double z) 
 	{
 		this.rotationX += x;
 		this.rotationY += y;
 		this.rotationZ += z;
 	}
 
-	public void setLocation(float x, float y, float z) 
+	public void setLocation(double x, double y, double z) 
 	{
 		this.x = x;
 		this.y = y;
 		this.z = z;
 	}
 
-	public void setRotation(float rotationX, float rotationY, float rotationZ) 
+	public void setRotation(double rotationX, double rotationY, double rotationZ) 
 	{
 		this.rotationX = rotationX;
 		this.rotationY = rotationY;
